remote queries: collapse the resolve action into resolveOnly on execute
The PAR bundle registered two Remote Queries actions, execute and resolve,
putting one product capability behind three independent FQN-scoped
authorization surfaces (the wf-actions-server service-account resource
allowlist, org Execution Policies, and the Agent PAR actions allowlist).
Two of the three denied resolve within one day on staging (2026-09-10 Test
Drive evidence), and every FQN-scoped customer policy would permanently
carry the allow-one-miss-the-other trap. Resolve is strictly weaker than
execute - execute runs the same matcher before SQL - and is not a
meaningful independent permission.

Retire the separate resolve action FQN and make resolution a mode of
execute, mirroring the AP-metadata change (dd-source PR #625) and the
worker change (dd-source PR #615):

- ExecuteInputs gains ResolveOnly (json:"resolveOnly"). When true, the
  action runs the existing resolve path unchanged: the same
  RemoteQueryResolve AgentSecure RPC, the same request mapping (exactly
  the integration and target), and the same resolve outputs.
- The resolve-mode input contract keeps the retired resolve action's
  side-effect-free structural guarantee: query, includeSchema,
  resultDelivery, and matchFingerprint are forbidden in that mode. The
  standalone action enforced this by decoding with
  DisallowUnknownFields; the merged action decodes tolerantly (execute
  inputs keep accepting unknown fields) and records the wire presence of
  the execute-only fields, and validateResolveOnlyInputs rejects that
  presence after decode and before the bridge client is ever created -
  the conditional equivalent of the old decode-time rejection, so a
  resolve-mode task cannot carry SQL or upload instructions even when
  their values are empty.
- When resolveOnly is false or absent, the execute path is unchanged:
  decode, validation, IPC request, and outputs keep today's behavior.
- Delete the resolve action registration and resolve.go; resolve_test.go
  coverage moves into execute_test.go as resolve-mode tests, including
  presence-based rejection cases and the single-action registration.

The AgentSecure IPC surface is unchanged: resolve mode still routes to
the existing RemoteQueryResolve RPC and the shared core proto. Remote
Queries is greenfield, so the retired FQN gets no compatibility path.
