+++ pkg/privateactionrunner/bundles/remoteaction/queries/execute.go
@@ -27,9 +27,10 @@ import (
 // emits it.
 const remoteQueryOperationProduceJSONPages = "produce_json_pages"
 
-// BridgeClient is the narrow AgentSecure gRPC client surface required by this bundle:
-// the streaming execute RPC and the unary side-effect-free resolve RPC. Bulk result
-// bytes never traverse AgentSecure, so there is no unary inline-result call.
+// BridgeClient is the narrow AgentSecure gRPC client surface required by this
+// bundle: the streaming execute RPC and the unary side-effect-free resolve RPC that
+// backs the execute action's resolveOnly mode. Bulk result bytes never traverse
+// AgentSecure, so there is no unary inline-result call.
 type BridgeClient interface {
 	RemoteQueryExecuteStream(ctx context.Context, in *pb.RemoteQueryExecuteRequest, opts ...grpc.CallOption) (grpc.ServerStreamingClient[pb.RemoteQueryExecuteChunk], error)
 	RemoteQueryResolve(ctx context.Context, in *pb.RemoteQueryResolveRequest, opts ...grpc.CallOption) (*pb.RemoteQueryResolveResponse, error)
@@ -38,6 +39,14 @@ type BridgeClient interface {
 // BridgeClientFactory returns an authenticated AgentSecure client over the local Agent IPC channel.
 type BridgeClientFactory func() (BridgeClient, error)
 
+// ExecuteAction runs the single Remote Queries AP action in one of two modes selected
+// by the resolveOnly input. The default execute mode dispatches customer SQL through
+// the streaming AgentSecure execute RPC with the backend-injected upload contract.
+// The resolveOnly mode is the side-effect-free target resolution: the Agent re-runs
+// the same integration matcher execute uses and answers the structured zero/one/many
+// outcome, with the opaque match fingerprint for the unique case. Resolve mode never
+// dispatches a query, never creates a page writer, and never touches upload
+// credentials — enforced by the mode's input contract and request mapping.
 type ExecuteAction struct {
 	newBridgeClient BridgeClientFactory
 }
@@ -51,9 +60,10 @@ func NewExecuteAction(newBridgeClient BridgeClientFactory) *ExecuteAction {
 // (authoritative run/task identity, artifact version, scoped upload instructions,
 // and effective limits), and the optional resolve-time matchFingerprint the Agent
 // revalidates before any SQL execution. The input carries no credentials: the org
-// API/application keys are read by the integration from Agent config, and the
-// session is identified solely by its upload id — there is no per-session upload
-// token.
+// API/application keys are read by the integration from Agent config, and the session
+// is identified solely by its upload id — there is no per-session upload token.
+// resolveOnly selects the side-effect-free resolution mode; when true the input is
+// target-only and the execute-only fields are forbidden (see validateResolveOnlyInputs).
 type ExecuteInputs struct {
 	Integration      string                `json:"integration"`
 	Target           TargetInputs          `json:"target"`
@@ -61,6 +71,54 @@ type ExecuteInputs struct {
 	IncludeSchema    bool                  `json:"includeSchema"`
 	ResultDelivery   *ResultDeliveryInputs `json:"resultDelivery"`
 	MatchFingerprint string                `json:"matchFingerprint"`
+	ResolveOnly      bool                  `json:"resolveOnly"`
+
+	// Presence of the execute-only fields on the wire. Resolve mode forbids those
+	// fields by presence — an empty query or a null resultDelivery is still a
+	// contract violation — so the decode records presence alongside the values.
+	querySet            bool
+	includeSchemaSet    bool
+	resultDeliverySet   bool
+	matchFingerprintSet bool
+}
+
+// UnmarshalJSON decodes the AP action input tolerantly — execute-mode inputs keep
+// accepting unknown fields, exactly as before — while recording whether the
+// execute-only fields were present on the wire, so validateResolveOnlyInputs can
+// distinguish "absent" from "present but zero-valued".
+func (i *ExecuteInputs) UnmarshalJSON(data []byte) error {
+	var raw map[string]json.RawMessage
+	if err := json.Unmarshal(data, &raw); err != nil {
+		return err
+	}
+
+	var wire struct {
+		Integration      string                `json:"integration"`
+		Target           TargetInputs          `json:"target"`
+		Query            string                `json:"query"`
+		IncludeSchema    bool                  `json:"includeSchema"`
+		ResultDelivery   *ResultDeliveryInputs `json:"resultDelivery"`
+		MatchFingerprint string                `json:"matchFingerprint"`
+		ResolveOnly      bool                  `json:"resolveOnly"`
+	}
+	if err := json.Unmarshal(data, &wire); err != nil {
+		return err
+	}
+
+	*i = ExecuteInputs{
+		Integration:      wire.Integration,
+		Target:           wire.Target,
+		Query:            wire.Query,
+		IncludeSchema:    wire.IncludeSchema,
+		ResultDelivery:   wire.ResultDelivery,
+		MatchFingerprint: wire.MatchFingerprint,
+		ResolveOnly:      wire.ResolveOnly,
+	}
+	_, i.querySet = raw["query"]
+	_, i.includeSchemaSet = raw["includeSchema"]
+	_, i.resultDeliverySet = raw["resultDelivery"]
+	_, i.matchFingerprintSet = raw["matchFingerprint"]
+	return nil
 }
 
 type ResultDeliveryInputs struct {
@@ -174,6 +232,32 @@ func validateDeliveryInputs(delivery *ResultDeliveryInputs) error {
 	return nil
 }
 
+// validateResolveOnlyInputs is the targeted strict validation of the resolveOnly
+// mode: a resolution dispatch carries exactly the integration and target, so the
+// execute-only fields — query, includeSchema, resultDelivery, and matchFingerprint —
+// are forbidden. Presence-based rejection is the conditional equivalent of the
+// structural guarantee the retired standalone resolve action enforced by decoding
+// with DisallowUnknownFields: a resolve-mode task cannot carry SQL, upload
+// instructions, or a fingerprint even when their values are empty, because the
+// fields themselves are the contract violation. The check runs after decode but
+// before the bridge client is ever created, so a malformed task never reaches the
+// Agent IPC.
+func validateResolveOnlyInputs(inputs ExecuteInputs) error {
+	if inputs.querySet {
+		return errors.New("resolveOnly input must not carry query")
+	}
+	if inputs.includeSchemaSet {
+		return errors.New("resolveOnly input must not carry includeSchema")
+	}
+	if inputs.resultDeliverySet {
+		return errors.New("resolveOnly input must not carry resultDelivery")
+	}
+	if inputs.matchFingerprintSet {
+		return errors.New("resolveOnly input must not carry matchFingerprint")
+	}
+	return nil
+}
+
 func (a *ExecuteAction) Run(
 	ctx context.Context,
 	task *types.Task,
@@ -193,7 +277,14 @@ func (a *ExecuteAction) Run(
 			"invalid remote query action inputs",
 		)
 	}
-	if err := validateDeliveryInputs(inputs.ResultDelivery); err != nil {
+	if inputs.ResolveOnly {
+		if err := validateResolveOnlyInputs(inputs); err != nil {
+			return nil, util.DefaultActionErrorWithDisplayError(
+				errors.New("invalid remote query action inputs"),
+				"invalid remote query action inputs",
+			)
+		}
+	} else if err := validateDeliveryInputs(inputs.ResultDelivery); err != nil {
 		return nil, util.DefaultActionErrorWithDisplayError(
 			errors.New("invalid remote query action inputs"),
 			"invalid remote query action inputs",
@@ -209,6 +300,18 @@ func (a *ExecuteAction) Run(
 	}
 	if client == nil {
 		return nil, util.DefaultActionError(errors.New("remote query action requires an AgentSecure client"))
+	}
+
+	if inputs.ResolveOnly {
+		resp, err := client.RemoteQueryResolve(ctx, remoteQueryResolveRequestFromInputs(inputs))
+		if err != nil {
+			return nil, util.DefaultActionErrorWithDisplayError(err, "remote query AgentSecure resolve RPC failed")
+		}
+		output, err := remoteQueryResolveOutputFromResponse(resp)
+		if err != nil {
+			return nil, util.DefaultActionErrorWithDisplayError(err, "remote query AgentSecure resolve RPC response was invalid")
+		}
+		return output, nil
 	}
 
 	stream, err := client.RemoteQueryExecuteStream(ctx, remoteQueryExecuteRequestFromInputs(inputs))
@@ -367,4 +470,50 @@ func remoteQueryErrorOutput(errEvent *pb.RemoteQueryStreamError) map[string]inte
 			"message": errEvent.GetMessage(),
 		},
 	}
+}
+
+// remoteQueryResolveRequestFromInputs maps the resolveOnly AP action input to the
+// credential-free AgentSecure resolve request: exactly the integration and the
+// target, mirroring execute's target mapping. The resolve request mapping carries
+// no query, no result delivery, and no fingerprint, so the side-effect-free
+// guarantee holds by construction of the mapping as well as of the input contract.
+func remoteQueryResolveRequestFromInputs(inputs ExecuteInputs) *pb.RemoteQueryResolveRequest {
+	return &pb.RemoteQueryResolveRequest{
+		Integration: inputs.Integration,
+		Target: &pb.RemoteQueryTarget{
+			Host:             inputs.Target.Host,
+			Port:             int32(inputs.Target.Port),
+			Dbname:           inputs.Target.DBName,
+			DatabaseInstance: inputs.Target.DatabaseInstance,
+		},
+	}
+}
+
+// remoteQueryResolveOutputFromResponse maps the typed AgentSecure resolve response
+// to the AP action output, mirroring execute's result-object conventions: exactly
+// {status}, plus matchFingerprint when the response carries one and the error
+// object when the response carries one. The mapping is deliberately opaque — an
+// unknown status or a matched response without a fingerprint passes through
+// unchanged so the dispatcher classifies well-formed contract violations itself
+// instead of the bundle collapsing them into a transport failure. Only a
+// structurally missing response or status fails closed.
+func remoteQueryResolveOutputFromResponse(resp *pb.RemoteQueryResolveResponse) (map[string]interface{}, error) {
+	if resp == nil {
+		return nil, errors.New("remote query resolve response missing")
+	}
+	status := resp.GetStatus()
+	if status == "" {
+		return nil, errors.New("remote query resolve response missing status")
+	}
+	output := map[string]interface{}{"status": status}
+	if fingerprint := resp.GetMatchFingerprint(); fingerprint != "" {
+		output["matchFingerprint"] = fingerprint
+	}
+	if code := resp.GetErrorCode(); code != "" {
+		output["error"] = map[string]interface{}{
+			"code":    code,
+			"message": resp.GetErrorMessage(),
+		}
+	}
+	return output, nil
 }
