+++ pkg/privateactionrunner/bundles/remoteaction/queries/execute_test.go
@@ -27,6 +27,10 @@ const (
 	testBaseURL  = "https://dd.datad0g.com/api/unstable/its-agent-intake"
 )
 
+// testFingerprint is the opaque resolve-time match fingerprint carried between the
+// resolve and execute modes.
+const testFingerprint = "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef"
+
 func resultDeliveryInputs() map[string]interface{} {
 	return map[string]interface{}{
 		"runId":           testRunID,
@@ -496,6 +500,430 @@ func TestExecuteActionSanitizesInputExtractionErrors(t *testing.T) {
 	assert.NotContains(t, err.Error(), "SELECT secret")
 }
 
+// TestBundleRegistersExecuteActionOnly proves the bundle registers the single
+// execute action: the standalone resolve action is retired and resolution is the
+// execute action's resolveOnly mode.
+func TestBundleRegistersExecuteActionOnly(t *testing.T) {
+	bundle := NewRemoteQueriesBundle()
+
+	require.NotNil(t, bundle.GetAction(ExecuteActionName))
+	assert.Nil(t, bundle.GetAction("resolve"))
+}
+
+// TestExecuteActionResolveOnlyUsesCredentialFreeAgentSecureRequestShape proves the
+// resolveOnly mode routes to the resolve RPC with exactly the integration and
+// target: no query, no result delivery, no fingerprint echo, the execute stream is
+// never opened, and the private credentials never reach the AgentSecure request.
+func TestExecuteActionResolveOnlyUsesCredentialFreeAgentSecureRequestShape(t *testing.T) {
+	client := &captureBridgeClient{resolveResp: &pb.RemoteQueryResolveResponse{
+		Status:           "matched",
+		MatchFingerprint: testFingerprint,
+	}}
+	action := NewExecuteAction(func() (BridgeClient, error) { return client, nil })
+
+	output, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+		"integration": "postgres",
+		"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "postgres"},
+	}), &privateconnection.PrivateCredentials{Tokens: []privateconnection.PrivateCredentialsToken{{Name: "password", Value: "secret-value"}}})
+
+	require.NoError(t, err)
+	require.NotNil(t, client.resolveRequest)
+	assert.Equal(t, "postgres", client.resolveRequest.GetIntegration())
+	assert.Equal(t, "localhost", client.resolveRequest.GetTarget().GetHost())
+	assert.Equal(t, int32(5432), client.resolveRequest.GetTarget().GetPort())
+	assert.Equal(t, "postgres", client.resolveRequest.GetTarget().GetDbname())
+	assert.Empty(t, client.resolveRequest.GetTarget().GetDatabaseInstance())
+
+	// The streaming execute RPC is never opened in resolve mode.
+	assert.Nil(t, client.request)
+
+	requestEvidence, err := json.Marshal(client.resolveRequest)
+	require.NoError(t, err)
+	assert.NotContains(t, string(requestEvidence), "secret-value")
+
+	out, ok := output.(map[string]interface{})
+	require.True(t, ok)
+	assert.Equal(t, map[string]interface{}{
+		"status":           "matched",
+		"matchFingerprint": testFingerprint,
+	}, out)
+}
+
+// TestExecuteActionResolveOnlyAcceptsDatabaseInstanceTarget proves the
+// managed-instance selector maps through the resolve request like execute's target
+// mapping.
+func TestExecuteActionResolveOnlyAcceptsDatabaseInstanceTarget(t *testing.T) {
+	client := &captureBridgeClient{resolveResp: &pb.RemoteQueryResolveResponse{
+		Status:           "matched",
+		MatchFingerprint: testFingerprint,
+	}}
+	action := NewExecuteAction(func() (BridgeClient, error) { return client, nil })
+
+	_, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+		"integration": "postgres",
+		"target":      map[string]interface{}{"database_instance": "Rq-Proof-A1-DB1"},
+	}), nil)
+
+	require.NoError(t, err)
+	require.NotNil(t, client.resolveRequest)
+	assert.Equal(t, "Rq-Proof-A1-DB1", client.resolveRequest.GetTarget().GetDatabaseInstance())
+	assert.Empty(t, client.resolveRequest.GetTarget().GetHost())
+}
+
+// TestExecuteActionExplicitFalseResolveOnlyRunsExecute proves an explicit
+// resolveOnly: false is the default execute mode: the resolve-mode forbidden-field
+// contract does not apply and the dispatch runs the streaming execute path with the
+// full execute input.
+func TestExecuteActionExplicitFalseResolveOnlyRunsExecute(t *testing.T) {
+	client := &captureBridgeClient{chunks: []*pb.RemoteQueryExecuteChunk{
+		finalEvent(0, validReceipt(), nil),
+		finalMarker(1),
+	}}
+	action := NewExecuteAction(func() (BridgeClient, error) { return client, nil })
+
+	output, err := action.Run(context.Background(), taskWithInputs(map[string]interface{}{
+		"integration":      "postgres",
+		"target":           map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "postgres"},
+		"query":            "SELECT city, country FROM cities ORDER BY city",
+		"resultDelivery":   resultDeliveryInputs(),
+		"matchFingerprint": testFingerprint,
+		"resolveOnly":      false,
+	}), nil)
+
+	require.NoError(t, err)
+	require.NotNil(t, client.request)
+	assert.Equal(t, "SELECT city, country FROM cities ORDER BY city", client.request.GetQuery())
+	assert.Equal(t, testFingerprint, client.request.GetMatchFingerprint())
+	assert.Equal(t, "SUCCEEDED", output.(map[string]interface{})["status"])
+}
+
+// TestExecuteActionResolveOnlyRejectsExecuteOnlyFieldsBeforeRPC proves the resolveOnly
+// mode is side-effect-free by construction: a resolve-mode input carrying any
+// execute-only field — query, includeSchema, resultDelivery, or even a fingerprint —
+// never creates the bridge client. Rejection is presence-based: the empty-string,
+// explicit-false, and null variants are contract violations too, mirroring the
+// retired standalone resolve action's DisallowUnknownFields structural guarantee.
+func TestExecuteActionResolveOnlyRejectsExecuteOnlyFieldsBeforeRPC(t *testing.T) {
+	executeOnlyInputs := []struct {
+		name  string
+		extra map[string]interface{}
+	}{
+		{
+			name:  "query",
+			extra: map[string]interface{}{"query": "SELECT secret FROM private_table"},
+		},
+		{
+			name:  "query empty but present",
+			extra: map[string]interface{}{"query": ""},
+		},
+		{
+			name:  "includeSchema",
+			extra: map[string]interface{}{"includeSchema": true},
+		},
+		{
+			name:  "includeSchema explicit false",
+			extra: map[string]interface{}{"includeSchema": false},
+		},
+		{
+			name:  "resultDelivery",
+			extra: map[string]interface{}{"resultDelivery": resultDeliveryInputs()},
+		},
+		{
+			name:  "resultDelivery null but present",
+			extra: map[string]interface{}{"resultDelivery": nil},
+		},
+		{
+			name:  "matchFingerprint",
+			extra: map[string]interface{}{"matchFingerprint": testFingerprint},
+		},
+		{
+			name:  "matchFingerprint empty but present",
+			extra: map[string]interface{}{"matchFingerprint": ""},
+		},
+	}
+
+	for _, tt := range executeOnlyInputs {
+		t.Run(tt.name, func(t *testing.T) {
+			action := NewExecuteAction(func() (BridgeClient, error) {
+				require.Fail(t, "bridge client should not be created for a resolveOnly input with execute-only fields")
+				return nil, nil
+			})
+
+			inputs := map[string]interface{}{
+				"integration": "postgres",
+				"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "postgres"},
+			}
+			for key, value := range tt.extra {
+				inputs[key] = value
+			}
+
+			_, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(inputs), nil)
+
+			require.Error(t, err)
+			var parErr util.PARError
+			require.ErrorAs(t, err, &parErr)
+			assert.Equal(t, "invalid remote query action inputs", parErr.Message)
+			assert.NotContains(t, err.Error(), "SELECT secret")
+			assert.NotContains(t, err.Error(), testFingerprint)
+		})
+	}
+}
+
+// TestExecuteActionResolveOnlyRejectsInvalidTargetBeforeRPC mirrors the execute-mode
+// target validation: malformed selectors never create the bridge client.
+func TestExecuteActionResolveOnlyRejectsInvalidTargetBeforeRPC(t *testing.T) {
+	tests := []struct {
+		name   string
+		target map[string]interface{}
+	}{
+		{name: "mixed selectors", target: map[string]interface{}{"database_instance": "rq-proof-a1-db1", "host": "localhost", "port": 5432, "dbname": "postgres"}},
+		{name: "partial tuple", target: map[string]interface{}{"host": "localhost", "dbname": "postgres"}},
+		{name: "whitespace instance", target: map[string]interface{}{"database_instance": " rq-proof-a1-db1 "}},
+		{name: "unknown credential field", target: map[string]interface{}{"database_instance": "rq-proof-a1-db1", "password": "secret-value"}},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			action := NewExecuteAction(func() (BridgeClient, error) {
+				require.Fail(t, "bridge client should not be created for an invalid resolveOnly target")
+				return nil, nil
+			})
+
+			_, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+				"integration": "postgres",
+				"target":      tt.target,
+			}), nil)
+
+			require.Error(t, err)
+			var parErr util.PARError
+			require.ErrorAs(t, err, &parErr)
+			assert.Equal(t, "invalid remote query action inputs", parErr.Message)
+			assert.NotContains(t, err.Error(), "secret-value")
+		})
+	}
+}
+
+// TestExecuteActionResolveOnlyPropagatesStructuredOutcomes proves the typed resolve
+// response maps to the AP output envelope: non-matched statuses carry the error object
+// mirroring the status, matched carries the fingerprint.
+func TestExecuteActionResolveOnlyPropagatesStructuredOutcomes(t *testing.T) {
+	t.Run("target not found", func(t *testing.T) {
+		client := &captureBridgeClient{resolveResp: &pb.RemoteQueryResolveResponse{
+			Status:       "target_not_found",
+			ErrorCode:    "target_not_found",
+			ErrorMessage: "no matching integration check found",
+		}}
+		action := NewExecuteAction(func() (BridgeClient, error) { return client, nil })
+
+		output, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+			"integration": "postgres",
+			"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "secret-db"},
+		}), nil)
+
+		require.NoError(t, err)
+		assert.Equal(t, map[string]interface{}{
+			"status": "target_not_found",
+			"error": map[string]interface{}{
+				"code":    "target_not_found",
+				"message": "no matching integration check found",
+			},
+		}, output)
+	})
+
+	t.Run("ambiguous", func(t *testing.T) {
+		client := &captureBridgeClient{resolveResp: &pb.RemoteQueryResolveResponse{
+			Status:       "ambiguous_target",
+			ErrorCode:    "ambiguous_target",
+			ErrorMessage: "multiple matching integration checks found",
+		}}
+		action := NewExecuteAction(func() (BridgeClient, error) { return client, nil })
+
+		output, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+			"integration": "postgres",
+			"target":      map[string]interface{}{"database_instance": "duplicate"},
+		}), nil)
+
+		require.NoError(t, err)
+		assert.Equal(t, map[string]interface{}{
+			"status": "ambiguous_target",
+			"error": map[string]interface{}{
+				"code":    "ambiguous_target",
+				"message": "multiple matching integration checks found",
+			},
+		}, output)
+	})
+
+	t.Run("resolution error", func(t *testing.T) {
+		client := &captureBridgeClient{resolveResp: &pb.RemoteQueryResolveResponse{
+			Status:       "resolution_error",
+			ErrorCode:    "resolution_error",
+			ErrorMessage: "remote queries resolve bridge is disabled",
+		}}
+		action := NewExecuteAction(func() (BridgeClient, error) { return client, nil })
+
+		output, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+			"integration": "postgres",
+			"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "postgres"},
+		}), nil)
+
+		require.NoError(t, err)
+		assert.Equal(t, map[string]interface{}{
+			"status": "resolution_error",
+			"error": map[string]interface{}{
+				"code":    "resolution_error",
+				"message": "remote queries resolve bridge is disabled",
+			},
+		}, output)
+	})
+}
+
+// TestExecuteActionResolveOnlyPassesThroughWellFormedContractViolations proves the
+// bundle maps resolve responses opaquely: a matched response without a fingerprint
+// and an unknown status pass through unchanged so the dispatcher classifies them
+// (resolution_error) instead of the bundle collapsing them into a transport failure.
+func TestExecuteActionResolveOnlyPassesThroughWellFormedContractViolations(t *testing.T) {
+	t.Run("matched without fingerprint", func(t *testing.T) {
+		client := &captureBridgeClient{resolveResp: &pb.RemoteQueryResolveResponse{Status: "matched"}}
+		action := NewExecuteAction(func() (BridgeClient, error) { return client, nil })
+
+		output, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+			"integration": "postgres",
+			"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "postgres"},
+		}), nil)
+
+		require.NoError(t, err)
+		assert.Equal(t, map[string]interface{}{"status": "matched"}, output)
+	})
+
+	t.Run("unknown status", func(t *testing.T) {
+		client := &captureBridgeClient{resolveResp: &pb.RemoteQueryResolveResponse{Status: "something_unexpected"}}
+		action := NewExecuteAction(func() (BridgeClient, error) { return client, nil })
+
+		output, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+			"integration": "postgres",
+			"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "postgres"},
+		}), nil)
+
+		require.NoError(t, err)
+		assert.Equal(t, map[string]interface{}{"status": "something_unexpected"}, output)
+	})
+}
+
+// TestExecuteActionResolveOnlyFailsClosedOnMalformedResponses proves structurally
+// invalid responses — a nil response or a missing status — surface as PAR action
+// errors with sanitized display messages.
+func TestExecuteActionResolveOnlyFailsClosedOnMalformedResponses(t *testing.T) {
+	tests := []struct {
+		name         string
+		resp         *pb.RemoteQueryResolveResponse
+		wantExternal string
+	}{
+		{name: "nil response", resp: nil, wantExternal: "remote query AgentSecure resolve RPC response was invalid"},
+		{name: "missing status", resp: &pb.RemoteQueryResolveResponse{}, wantExternal: "remote query AgentSecure resolve RPC response was invalid"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			client := &captureBridgeClient{resolveResp: tt.resp}
+			action := NewExecuteAction(func() (BridgeClient, error) { return client, nil })
+
+			_, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+				"integration": "postgres",
+				"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "postgres"},
+			}), nil)
+
+			require.Error(t, err)
+			var parErr util.PARError
+			require.ErrorAs(t, err, &parErr)
+			assert.Equal(t, tt.wantExternal, parErr.ExternalMessage)
+		})
+	}
+}
+
+// TestExecuteActionResolveOnlySurfacesTransportFailure proves an AgentSecure transport
+// failure on the resolve RPC surfaces as a sanitized PAR action error.
+func TestExecuteActionResolveOnlySurfacesTransportFailure(t *testing.T) {
+	client := &captureBridgeClient{err: assert.AnError}
+	action := NewExecuteAction(func() (BridgeClient, error) { return client, nil })
+
+	_, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+		"integration": "postgres",
+		"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "postgres"},
+	}), nil)
+
+	require.Error(t, err)
+	var parErr util.PARError
+	require.ErrorAs(t, err, &parErr)
+	assert.Equal(t, "remote query AgentSecure resolve RPC failed", parErr.ExternalMessage)
+}
+
+// TestExecuteActionResolveOnlyRequiresBridgeClient exercises the shared bridge client
+// guards through the resolveOnly mode.
+func TestExecuteActionResolveOnlyRequiresBridgeClient(t *testing.T) {
+	t.Run("missing factory", func(t *testing.T) {
+		var action *ExecuteAction
+
+		_, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+			"integration": "postgres",
+			"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "postgres"},
+		}), nil)
+
+		require.Error(t, err)
+		var parErr util.PARError
+		require.ErrorAs(t, err, &parErr)
+		assert.Equal(t, "remote query action requires an Agent IPC client", parErr.Message)
+	})
+
+	t.Run("client creation failure", func(t *testing.T) {
+		action := NewExecuteAction(func() (BridgeClient, error) { return nil, assert.AnError })
+
+		_, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+			"integration": "postgres",
+			"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "postgres"},
+		}), nil)
+
+		require.Error(t, err)
+		var parErr util.PARError
+		require.ErrorAs(t, err, &parErr)
+		assert.Equal(t, "remote query action could not create an Agent IPC client", parErr.ExternalMessage)
+	})
+
+	t.Run("nil client", func(t *testing.T) {
+		action := NewExecuteAction(func() (BridgeClient, error) { return nil, nil })
+
+		_, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+			"integration": "postgres",
+			"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "postgres"},
+		}), nil)
+
+		require.Error(t, err)
+		var parErr util.PARError
+		require.ErrorAs(t, err, &parErr)
+		assert.Equal(t, "remote query action requires an AgentSecure client", parErr.Message)
+	})
+}
+
+// TestExecuteActionResolveOnlySanitizesInputExtractionErrors proves a malformed
+// resolve-mode task input fails closed before the bridge client with a constant,
+// sanitized error.
+func TestExecuteActionResolveOnlySanitizesInputExtractionErrors(t *testing.T) {
+	action := NewExecuteAction(func() (BridgeClient, error) {
+		require.Fail(t, "bridge client should not be created for invalid inputs")
+		return nil, nil
+	})
+
+	_, err := action.Run(context.Background(), resolveOnlyTaskWithInputs(map[string]interface{}{
+		"integration": "postgres",
+		"target":      map[string]interface{}{"host": "localhost", "port": 5432, "dbname": "secret-db"},
+		"bad":         make(chan struct{}),
+	}), nil)
+
+	require.Error(t, err)
+	var parErr util.PARError
+	require.ErrorAs(t, err, &parErr)
+	assert.Equal(t, "invalid remote query action inputs", parErr.Message)
+	assert.Equal(t, "invalid remote query action inputs", parErr.ExternalMessage)
+	assert.NotContains(t, err.Error(), "secret-db")
+}
+
 // TestRemoteQueryExecuteOutputStaysUnderActionPlatformLimit proves the receipt-only
 // output is bounded by construction: with no inline result-byte path the AP artifact
 // stays tiny even for multi-page runs.
@@ -546,6 +974,14 @@ func taskWithInputs(inputs map[string]interface{}) *types.Task {
 		Inputs:   inputs,
 	}
 	return task
+}
+
+// resolveOnlyTaskWithInputs builds an execute task dispatching the resolveOnly mode.
+// The action name stays execute — resolution is a mode of the execute action, not a
+// separate action.
+func resolveOnlyTaskWithInputs(inputs map[string]interface{}) *types.Task {
+	inputs["resolveOnly"] = true
+	return taskWithInputs(inputs)
 }
 
 type captureBridgeClient struct {
